requirements.txt
dotenv
binance-futures-connector
